# openapi.yaml
openapi: 3.0.0
info:
  title: AIQuiz API
  description: Complete API documentation for the AIQuiz educational platform with AI-powered question generation and RAG-enabled document processing
  version: 1.0.0
  contact:
    name: AIQuiz Team
servers:
  - url: /api
    description: Development server
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/manager/auth/login
  schemas:
...
    Topic:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        subject:
          type: string
        subtopics:
          type: array
          items:
            $ref: '#/components/schemas/Subtopic'
    Subtopic:
      type: object
      properties:
        _id:
          type: string
        title:
          type: string
        description:
          type: string
        topic:
          type: string
        files:
          type: array
          items:
            $ref: '#/components/schemas/File'
...
    File:
      type: object
      properties:
        _id:
          type: string
        fileName:
          type: string
        originalName:
          type: string
        mimeType:
          type: string
        size:
          type: number
        fileType:
          type: string
          enum: [document, image, video, other]
        subtopic:
          type: string
        ragProcessed:
          type: boolean
        ragDocumentId:
          type: string
        ragStats:
          type: object
...
paths:
...
  /manager/subjects/{id}/topics/{topicId}:
    get:
      summary: Get topic details
      tags:
        - Topic Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Topic details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Topic'

    put:
      summary: Update topic
      tags:
        - Topic Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '200':
          description: Topic updated successfully

    delete:
      summary: Delete topic
      tags:
        - Topic Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Topic deleted successfully

  # Subtopic Management
  /manager/subjects/{id}/topics/{topicId}/subtopics:
    get:
      summary: Get subtopics for topic
      tags:
        - Subtopic Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: List of subtopics
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Subtopic'

    post:
      summary: Create new subtopic
      tags:
        - Subtopic Management
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
        - name: topicId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
              properties:
                title:
                  type: string
                description:
                  type: string
      responses:
        '201':
          description: Subtopic created successfully
...